Split everything else out, make folder for lib code
