+++ Wangview/Tileset.py
@@ -1,0 +1,12 @@
+
+# coding: utf-8
+
+# In[ ]:
+
+from collections import namedtuple
+
+
+# In[ ]:
+
+Tileset = namedtuple('Tileset',['filename','offset','width','height','clique'])
+
