+++ WangviewIPython.py
@@ -1,0 +1,23 @@
+
+# coding: utf-8
+
+# In[ ]:
+
+get_ipython().magic('load_ext autoreload')
+
+
+# In[ ]:
+
+get_ipython().magic('autoreload 2')
+
+
+# In[ ]:
+
+from Wangview.Display import Display
+
+
+# In[ ]:
+
+w = Display('../Wangscape/Wangscape/example3/output')
+w.run()
+
